+++ src/AbatabLogging/WriteFile.cs
@@ -1,0 +1,12 @@
+﻿/* ========================================================================================================
+ * AbatabLogging.WriteFile.cs: Writes content to a local file.
+ * b220922.121116
+ * https://github.com/spectrum-health-systems/Abatab/blob/main/Documentation/Sourcecode/Sourcecode.md
+ * ===================================================================================================== */
+
+namespace AbatabLogging
+{
+    internal class WriteFile
+    {
+    }
+}