+++ src/AbatabOptionObject/Verify.cs
@@ -1,0 +1,18 @@
+﻿/* ========================================================================================================
+ * AbatabOptionObject.Verify.cs: Verify OptionObject2015 components.
+ * b220922.121445
+ * https://github.com/spectrum-health-systems/Abatab/blob/main/Documentation/Sourcecode/Sourcecode.md
+ * ===================================================================================================== */
+
+using NTST.ScriptLinkService.Objects;
+
+namespace AbatabOptionObject
+{
+    public static class Verify
+    {
+        public static bool NotModified(OptionObject2015 sentOptObj, OptionObject2015 alternateOptObj)
+        {
+            return alternateOptObj == sentOptObj;
+        }
+    }
+}