+++ src/AbatabRoundhouse/Roundhouse.cs
@@ -1,0 +1,26 @@
+﻿/* ========================================================================================================
+ * AbatabRoundhouse.Roundhouse.cs: The roundhouse for Abatab.
+ * b220922.120913
+ * https://github.com/spectrum-health-systems/Abatab/blob/main/Documentation/Sourcecode/Sourcecode.md
+ * ===================================================================================================== */
+
+namespace AbatabRoundhouse
+{
+    public class Roundhouse
+    {
+        public static void TestA()
+        {
+
+        }
+
+        public static void TestB()
+        {
+
+        }
+
+        public static void TestC()
+        {
+
+        }
+    }
+}