+++ Documentation/CHANGELOG.md
@@ -17,11 +17,39 @@
       - Project logo should be located at "./.github/Logos/ProjectLogo.png".
       - Short description of the project.
   -->
-  ![PROJECT-LOGO][PROJECT-LOGO]
+  ![REPOSITORY-LOGO][REPOSITORY-LOGO]
 
 </div>
 
-# vMajor.Minor.Patch-YYMMDD.HHMMSS-<descriptor>
+# v0.1.x - v0.9.x-devbuild
+
+## v0.4.1
+`INFO` Comment cleanup
+
+## v0.4.0
+`NEW` AbatabData.csproj
+`NEW` AbatabData.SessionData.cs
+
+## v0.3.1
+`NEW` AbatabOptionObject.csproj
+`NEW` AbatabOptionObject.Verify()
+
+## v0.3.0
+`NEW` AbatabLogging.csproj
+`NEW` AbatabOptionObject.csproj
+`NEW` AbatabSession.csproj
+
+## v0.2.2
+`INFO` Code/comment cleanup
+
+## v0.2.1
+`INFO` Code/comment cleanup
+
+## v0.2.0
+`INFO` Comment cleanup
+
+## v0.1.0
+`INFO` Initial framework commit
 
 <!-- BREADCRUMBS
      - Documentation breadcrumbs. This should also be at the start of the documentation as well.
@@ -36,4 +64,4 @@
      These reference links should be standard across all project documentation.
 -->
 [REPOSITORY-URL]: https://github.com/spectrum-health-systems/Abatab
-[PROJECT-LOGO]: ../.github/Logos/ProjectLogo.png
+[REPOSITORY-LOGO]: ../.github/Logos/ProjectLogo.png